Split the implementation of putchar from printf
The implementation of `putchar` may be various in different devices.

We use an platform-independent function `platform_putchar` to replace
the previous codes in print.cpp. So we can implement our own `putchar`
function in src/hw/<specific_device>/port.c instead of placing all
implementation in print.cpp.

In case that we need to initialize the device before using the platform
specific `putchar`, we also provide `platform_putc_init` function.