Port AMT onto RPi4B
In this commit, we add `raspi4` as a new build platform and create
a new category in `src/hw/` for raspi4-related implementations.

Since the bare metal way of reading/writing on SD card is still
working in progress[1], only the demos that are not related to file
system can be executed.

We also provides part of the functions in <string.h> to avoid
unalignment access from using toolchain-provided functions.

[1]: https://www.raspberrypi.org/forums/viewtopic.php?f=72&t=308089